+++ srcd/contrib/docker/bootstrap.py
@@ -3,7 +3,7 @@ import os
 from flask import g
 from flask_migrate import upgrade as db_upgrade
 
-from superset import app, conf, db, security_manager
+from superset import app, appbuilder, conf, db, security_manager
 from superset.connectors.connector_registry import ConnectorRegistry
 from superset.models import core as models
 from superset.models.user_attributes import UserAttribute
@@ -27,7 +27,9 @@ def get_or_create_datasource(name, uri, **kwargs):
 
 def create_datasource_tables(dbobj, schema):
     TBL = ConnectorRegistry.sources['table']
-    for table in dbobj.all_table_names_in_schema(schema):
+    for ds in dbobj.get_all_table_names_in_schema(schema):
+        table = ds.table
+
         # table_name should match the one in the datasource for fetch_metadata to work
         if db.session.query(TBL).filter_by(table_name=table).first():
             continue
@@ -96,6 +98,7 @@ def bootstrap():
                                                os.environ['ADMIN_PASSWORD'])
         # Create default roles and permissions
         utils.get_or_create_main_db()
+        appbuilder.add_permissions(update_perms=True)
         security_manager.sync_role_definitions()
 
         # set admin user as a current user