+++ superset/superset/assets/src/chart/ChartRenderer.jsx
@@ -20,17 +20,16 @@ import dompurify from 'dompurify';
 import { snakeCase } from 'lodash';
 import PropTypes from 'prop-types';
 import React from 'react';
-import { ChartProps, SuperChart } from '@superset-ui/chart';
+import { SuperChart } from '@superset-ui/chart';
 import { Tooltip } from 'react-bootstrap';
 import { Logger, LOG_ACTIONS_RENDER_CHART } from '../logger/LogUtils';
-import transformBigNumber from './transformBigNumber';
 
 const propTypes = {
   annotationData: PropTypes.object,
   actions: PropTypes.object,
   chartId: PropTypes.number.isRequired,
   datasource: PropTypes.object.isRequired,
-  filters: PropTypes.object,
+  initialValues: PropTypes.object,
   formData: PropTypes.object.isRequired,
   height: PropTypes.number,
   width: PropTypes.number,
@@ -51,7 +50,7 @@ const BLANK = {};
 
 const defaultProps = {
   addFilter: () => BLANK,
-  filters: BLANK,
+  initialValues: BLANK,
   setControlValue() {},
   triggerRender: false,
 };
@@ -61,14 +60,12 @@ class ChartRenderer extends React.Component {
     super(props);
     this.state = {};
 
-    this.createChartProps = ChartProps.createSelector();
     this.hasQueryResponseChange = false;
 
     this.setTooltip = this.setTooltip.bind(this);
     this.handleAddFilter = this.handleAddFilter.bind(this);
     this.handleRenderSuccess = this.handleRenderSuccess.bind(this);
     this.handleRenderFailure = this.handleRenderFailure.bind(this);
-    this.preTransformProps = this.preTransformProps.bind(this);
   }
 
   shouldComponentUpdate(nextProps, nextState) {
@@ -87,7 +84,8 @@ class ChartRenderer extends React.Component {
         nextProps.height !== this.props.height ||
         nextProps.width !== this.props.width ||
         nextState.tooltip !== this.state.tooltip ||
-        nextProps.triggerRender) {
+        nextProps.triggerRender ||
+        nextProps.formData.color_scheme !== this.props.formData.color_scheme) {
         return true;
       }
     }
@@ -96,33 +94,6 @@ class ChartRenderer extends React.Component {
 
   setTooltip(tooltip) {
     this.setState({ tooltip });
-  }
-
-  prepareChartProps() {
-    const {
-      width,
-      height,
-      annotationData,
-      datasource,
-      filters,
-      formData,
-      queryResponse,
-      setControlValue,
-    } = this.props;
-
-    return this.createChartProps({
-      width,
-      height,
-      annotationData,
-      datasource,
-      filters,
-      formData,
-      onAddFilter: this.handleAddFilter,
-      onError: this.handleRenderFailure,
-      payload: queryResponse,
-      setControlValue,
-      setTooltip: this.setTooltip,
-    });
   }
 
   handleAddFilter(col, vals, merge = true, refresh = true) {
@@ -166,18 +137,6 @@ class ChartRenderer extends React.Component {
     }
   }
 
-  preTransformProps(chartProps) {
-    const payload = chartProps.payload;
-    const data = transformBigNumber(payload.data);
-    return new ChartProps({
-      ...chartProps,
-      payload: {
-        ...payload,
-        data,
-      },
-    });
-  }
-
   renderTooltip() {
     const { tooltip } = this.state;
     if (tooltip && tooltip.content) {
@@ -210,10 +169,23 @@ class ChartRenderer extends React.Component {
       chartId,
     } = this.props;
 
-    const isLoading = chartStatus === 'loading';
+    // Skip chart rendering
+    if (chartStatus === 'loading' || !!chartAlert || chartStatus === null) {
+      return null;
+    }
 
-    const skipChartRendering = isLoading || !!chartAlert;
     this.renderStartTime = Logger.getTimestamp();
+
+    const {
+      width,
+      height,
+      annotationData,
+      datasource,
+      initialValues,
+      formData,
+      queryResponse,
+      setControlValue,
+    } = this.props;
 
     return (
       <React.Fragment>
@@ -222,8 +194,17 @@ class ChartRenderer extends React.Component {
           id={`chart-id-${chartId}`}
           className={`${snakeCase(vizType)}`}
           chartType={vizType}
-          chartProps={skipChartRendering ? null : this.prepareChartProps()}
-          preTransformProps={this.preTransformProps}
+          width={width}
+          height={height}
+          annotationData={annotationData}
+          datasource={datasource}
+          filters={initialValues}
+          formData={formData}
+          payload={queryResponse}
+          onAddFilter={this.handleAddFilter}
+          onError={this.handleRenderFailure}
+          setControlValue={setControlValue}
+          setTooltip={this.setTooltip}
           onRenderSuccess={this.handleRenderSuccess}
           onRenderFailure={this.handleRenderFailure}
         />