+++ superset/superset/assets/src/dashboard/containers/DashboardHeader.jsx
@@ -24,9 +24,10 @@ import isDashboardLoading from '../util/isDashboardLoading';
 
 import {
   setEditMode,
-  toggleBuilderPane,
+  showBuilderPane,
   fetchFaveStar,
   saveFaveStar,
+  savePublished,
   fetchCharts,
   startPeriodicRender,
   updateCss,
@@ -71,14 +72,17 @@ function mapStateToProps({
     expandedSlices: dashboardState.expandedSlices,
     refreshFrequency: dashboardState.refreshFrequency,
     css: dashboardState.css,
+    colorNamespace: dashboardState.colorNamespace,
+    colorScheme: dashboardState.colorScheme,
     charts,
     userId: dashboardInfo.userId,
     isStarred: !!dashboardState.isStarred,
+    isPublished: !!dashboardState.isPublished,
     isLoading: isDashboardLoading(charts),
     hasUnsavedChanges: !!dashboardState.hasUnsavedChanges,
     maxUndoHistoryExceeded: !!dashboardState.maxUndoHistoryExceeded,
     editMode: !!dashboardState.editMode,
-    showBuilderPane: !!dashboardState.showBuilderPane,
+    builderPaneType: dashboardState.builderPaneType,
   };
 }
 
@@ -91,9 +95,10 @@ function mapDispatchToProps(dispatch) {
       onUndo: undoLayoutAction,
       onRedo: redoLayoutAction,
       setEditMode,
-      toggleBuilderPane,
+      showBuilderPane,
       fetchFaveStar,
       saveFaveStar,
+      savePublished,
       fetchCharts,
       startPeriodicRender,
       updateDashboardTitle,