+++ superset/superset/assets/src/visualizations/deckgl/Multi/Multi.jsx
@@ -17,6 +17,7 @@
  * under the License.
  */
 import React from 'react';
+import _ from 'lodash';
 import PropTypes from 'prop-types';
 import { SupersetClient } from '@superset-ui/connection';
 
@@ -29,12 +30,21 @@ const propTypes = {
   payload: PropTypes.object.isRequired,
   setControlValue: PropTypes.func.isRequired,
   viewport: PropTypes.object.isRequired,
+  onAddFilter: PropTypes.func,
+  setTooltip: PropTypes.func,
+  onSelect: PropTypes.func,
+};
+const defaultProps = {
+  onAddFilter() {},
+  setTooltip() {},
+  onSelect() {},
 };
 
 class DeckMulti extends React.PureComponent {
   constructor(props) {
     super(props);
     this.state = { subSlicesLayers: {} };
+    this.onViewportChange = this.onViewportChange.bind(this);
   }
 
   componentDidMount() {
@@ -44,11 +54,18 @@ class DeckMulti extends React.PureComponent {
 
   componentWillReceiveProps(nextProps) {
     const { formData, payload } = nextProps;
-    this.loadLayers(formData, payload);
+    const hasChanges = !_.isEqual(this.props.formData.deck_slices, nextProps.formData.deck_slices);
+    if (hasChanges) {
+      this.loadLayers(formData, payload);
+    }
   }
 
-  loadLayers(formData, payload) {
-    this.setState({ subSlicesLayers: {} });
+  onViewportChange(viewport) {
+    this.setState({ viewport });
+  }
+
+  loadLayers(formData, payload, viewport) {
+    this.setState({ subSlicesLayers: {}, viewport });
     payload.data.slices.forEach((subslice) => {
       // Filters applied to multi_deck are passed down to underlying charts
       // note that dashboard contextual information (filter_immune_slices and such) aren't
@@ -56,7 +73,7 @@ class DeckMulti extends React.PureComponent {
       const filters = [
         ...(subslice.form_data.filters || []),
         ...(formData.filters || []),
-        ...(formData.extraFilters || []),
+        ...(formData.extra_filters || []),
       ];
       const subsliceCopy = {
         ...subslice,
@@ -70,9 +87,13 @@ class DeckMulti extends React.PureComponent {
           endpoint: getExploreLongUrl(subsliceCopy.form_data, 'json'),
         })
         .then(({ json }) => {
-          const layer = layerGenerators[subsliceCopy.form_data.vizType](
+          const layer = layerGenerators[subsliceCopy.form_data.viz_type](
             subsliceCopy.form_data,
             json,
+            this.props.onAddFilter,
+            this.props.setTooltip,
+            [],
+            this.props.onSelect,
           );
           this.setState({
             subSlicesLayers: {
@@ -86,7 +107,7 @@ class DeckMulti extends React.PureComponent {
   }
 
   render() {
-    const { payload, viewport, formData, setControlValue } = this.props;
+    const { payload, formData, setControlValue } = this.props;
     const { subSlicesLayers } = this.state;
 
     const layers = Object.values(subSlicesLayers);
@@ -94,9 +115,10 @@ class DeckMulti extends React.PureComponent {
     return (
       <DeckGLContainer
         mapboxApiAccessToken={payload.data.mapboxApiKey}
-        viewport={viewport}
+        viewport={this.state.viewport || this.props.viewport}
+        onViewportChange={this.onViewportChange}
         layers={layers}
-        mapStyle={formData.mapboxStyle}
+        mapStyle={formData.mapbox_style}
         setControlValue={setControlValue}
       />
     );
@@ -104,5 +126,6 @@ class DeckMulti extends React.PureComponent {
 }
 
 DeckMulti.propTypes = propTypes;
+DeckMulti.defaultProps = defaultProps;
 
 export default DeckMulti;