+++ superset/superset/views/__init__.py
@@ -15,11 +15,13 @@
 # specific language governing permissions and limitations
 # under the License.
 from . import base  # noqa
-from . import api # noqa
+from . import api  # noqa
 from . import core  # noqa
 from . import sql_lab  # noqa
-from . import dashboard # noqa
-from . import annotations # noqa
-from . import datasource # noqa
-from . import schedules # noqa
-from . import tags # noqa
+from . import dashboard  # noqa
+from . import annotations  # noqa
+from . import datasource  # noqa
+from . import schedules  # noqa
+from . import tags  # noqa
+from .log import views  # noqa
+from .log import api as log_api  # noqa