+++ CHANGELOG.md
@@ -7,6 +7,10 @@
     Click to see more.
   </summary>
 
+### Added
+
+- Superset updated to v0.34.0rc1 which brings many improvements and bug fixes ([#250](https://github.com/src-d/sourced-ui/issues/250))
+
 </details>
 
 ## [v0.5.0](https://github.com/src-d/sourced-ui/releases/tag/v0.5.0) - 2019-08-20